requirements.txt
torch
transformers
datasets
accelerate
peft
openai
aiohttp
tenacity
numpy
pandas
python-dotenv
pydantic
tqdm
tabulate
typing-extensions
